+++ src/args.rs
@@ -72,6 +72,10 @@ pub struct CheckArgs {
     #[arg(long, default_value = DEFAULT_PATH_DICTS)]
     pub path_dicts: PathBuf,
 
+    /// Path to a directory containing files with list of words to add per language (files are `*.dic`, eg: `en_US.dic`, with one word per line)
+    #[arg(long)]
+    pub path_words: Option<PathBuf>,
+
     /// Language used to check source strings
     #[arg(long, default_value = DEFAULT_LANG_ID)]
     pub lang_id: String,